Move data classes to separate files
