+++ app/src/main/java/com/marekmacko/kotlinapp/data/City.kt
@@ -1,0 +1,5 @@
+package com.marekmacko.kotlinapp.data
+
+data class City(val id: Long, val name: String,
+                val coord: Coordinates, val country: String,
+                val population: Int)
