+++ app/src/main/java/com/marekmacko/kotlinapp/data/Coordinates.kt
@@ -1,0 +1,3 @@
+package com.marekmacko.kotlinapp.data
+
+data class Coordinates(val lon: Float, val lat: Float)
